Move match function to Flow
